// generated/nextjs/authentication/types.ts
export interface AuthenticationStorage {
  readonly get: (relation: string, key: string) => Promise<Record<string, unknown> | null>;
  readonly put: (relation: string, key: string, value: Record<string, unknown>) => Promise<void>;
  readonly delete: (relation: string, key: string) => Promise<boolean>;
  readonly find: (relation: string, filter?: Record<string, unknown>) => Promise<readonly Record<string, unknown>[]>;
}

export interface AuthenticationRegisterInput {
  readonly user: string;
  readonly provider: string;
  readonly credentials: string;
}

export interface AuthenticationRegisterOutputOk {
  readonly variant: 'ok';
  readonly user: string;
}

export interface AuthenticationRegisterOutputExists {
  readonly variant: 'exists';
  readonly message: string;
}

export type AuthenticationRegisterOutput = AuthenticationRegisterOutputOk | AuthenticationRegisterOutputExists;

export interface AuthenticationLoginInput {
  readonly user: string;
  readonly credentials: string;
}

export interface AuthenticationLoginOutputOk {
  readonly variant: 'ok';
  readonly token: string;
}

export interface AuthenticationLoginOutputInvalid {
  readonly variant: 'invalid';
  readonly message: string;
}

export type AuthenticationLoginOutput = AuthenticationLoginOutputOk | AuthenticationLoginOutputInvalid;

export interface AuthenticationLogoutInput {
  readonly user: string;
}

export interface AuthenticationLogoutOutputOk {
  readonly variant: 'ok';
  readonly user: string;
}

export interface AuthenticationLogoutOutputNotfound {
  readonly variant: 'notfound';
  readonly message: string;
}

export type AuthenticationLogoutOutput = AuthenticationLogoutOutputOk | AuthenticationLogoutOutputNotfound;

export interface AuthenticationAuthenticateInput {
  readonly token: string;
}

export interface AuthenticationAuthenticateOutputOk {
  readonly variant: 'ok';
  readonly user: string;
}

export interface AuthenticationAuthenticateOutputInvalid {
  readonly variant: 'invalid';
  readonly message: string;
}

export type AuthenticationAuthenticateOutput = AuthenticationAuthenticateOutputOk | AuthenticationAuthenticateOutputInvalid;

export interface AuthenticationResetPasswordInput {
  readonly user: string;
  readonly newCredentials: string;
}

export interface AuthenticationResetPasswordOutputOk {
  readonly variant: 'ok';
  readonly user: string;
}

export interface AuthenticationResetPasswordOutputNotfound {
  readonly variant: 'notfound';
  readonly message: string;
}

export type AuthenticationResetPasswordOutput = AuthenticationResetPasswordOutputOk | AuthenticationResetPasswordOutputNotfound;